Remove unused arcReducer
Move the used functions to lib/proposalHelpers.ts

This removes all duplicate interfaces and loads them all from the client library. Except for IRedemptionState which doesnt exist in the client library
