+++ src/components/Dao/ProposalSchemeCard.tsx
@@ -7,7 +7,7 @@ import withSubscription, { ISubscriptionProps } from "components/Shared/withSubs
 import { humanProposalTitle, schemeName } from "lib/util";
 import * as React from "react";
 import { Link } from "react-router-dom";
-import { closingTime } from "reducers/arcReducer";
+import { closingTime } from "lib/proposalHelpers";
 import { combineLatest } from "rxjs";
 import TrainingTooltip from "components/Shared/TrainingTooltip";
 import * as css from "./SchemeCard.scss";
