+++ src/components/Proposal/ProposalDetailsPage.tsx
@@ -10,8 +10,7 @@ import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
 
 import { Link, RouteComponentProps } from "react-router-dom";
-import { proposalEnded } from "reducers/arcReducer";
-import { closingTime } from "reducers/arcReducer";
+import { closingTime, proposalEnded } from "lib/proposalHelpers";
 import TagsSelector from "components/Proposal/Create/SchemeForms/TagsSelector";
 import SocialShareModal from "../Shared/SocialShareModal";
 import ActionButton from "./ActionButton";
