+++ src/components/Proposal/ProposalHistoryRow.tsx
@@ -9,8 +9,7 @@ import * as React from "react";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
 import { IRootState } from "reducers";
-import { proposalFailed, proposalPassed } from "reducers/arcReducer";
-import { closingTime } from "reducers/arcReducer";
+import { closingTime, proposalFailed, proposalPassed } from "lib/proposalHelpers";
 import { IProfileState } from "reducers/profilesReducer";
 import { combineLatest, of } from "rxjs";
 import StakeGraph from "./Staking/StakeGraph";
