+++ src/components/Proposal/ProposalStatus.tsx
@@ -1,7 +1,7 @@
 import { IProposalStage, IProposalState } from "@daostack/client";
 import * as classNames from "classnames";
 import * as React from "react";
-import { proposalExpired, proposalFailed, proposalPassed } from "reducers/arcReducer";
+import { proposalExpired, proposalFailed, proposalPassed } from "lib/proposalHelpers";
 import * as css from "./ProposalStatus.scss";
 
 export default class ProposalStatus extends React.Component<IProps, null> {
