+++ src/reducers/index.ts
@@ -1,13 +1,11 @@
 import { routerReducer } from "react-router-redux";
 import { combineReducers } from "redux";
-import arcReducer, { IArcState } from "./arcReducer";
 import { INotificationsState, notificationsReducer } from "./notifications";
 import profilesReducer, { IProfilesState } from "./profilesReducer";
 import uiReducer, { IUIState } from "./uiReducer";
 import web3Reducer, { IWeb3State } from "./web3Reducer";
 
 export interface IRootState {
-  arc: IArcState;
   notifications: INotificationsState;
   profiles: IProfilesState;
   router: any;
@@ -16,7 +14,6 @@ export interface IRootState {
 }
 
 const reducers = {
-  arc: arcReducer,
   notifications: notificationsReducer,
   profiles: profilesReducer,
   router: routerReducer,
